+++ wsidicomizer/__init__.py
@@ -14,6 +14,6 @@
 
 from wsidicomizer.wsidicomizer import WsiDicomizer
 
-__version__ = "0.22.0"
+__version__ = "0.22.1"
 
 __all__ = ["WsiDicomizer"]